requirements.txt
numpy>=1.18.0
spatialmath-python>=0.8.7
roboticstoolbox-python>=0.8
pyyaml